move DHCPHeader to eth package
